backend/requirements.txt
fastapi
uvicorn
gunicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
alembic
python-multipart